test: RedisJson 테스트
알람 구현을 위한 RedisJson 형식 테스트

+++ portfolio-service/src/main/java/com/pda/portfolio_service/redis/RebalancingData.java
@@ -36,7 +36,7 @@ public class RebalancingData {
     }
     @Data
     public static class WeightData {
-        private double afterWeights;
-        private double beforeWeights;
+        private float afterWeights;
+        private float beforeWeights;
     }
 }
